// src/types/calendar.ts
export enum CalendarLegendItemType {
  None = "None",
  Today = "Today",
  Selected = "Selected",
  Working = "Working",
  NonWorking = "NonWorking",
  Type01 = "Type01",
  Type02 = "Type02",
  Type03 = "Type03",
  Type04 = "Type04",
  Type05 = "Type05",
  Type06 = "Type06",
  Type07 = "Type07",
  Type08 = "Type08",
  Type09 = "Type09",
  Type10 = "Type10",
  Type11 = "Type11",
  Type12 = "Type12",
  Type13 = "Type13",
  Type14 = "Type14",
  Type15 = "Type15",
  Type16 = "Type16",
  Type17 = "Type17",
  Type18 = "Type18",
  Type19 = "Type19",
  Type20 = "Type20",
}